feat: myPage script 추가 및 util 나누기
